Extrait les résultats de recherche dans un composant SearchResults
